feat: Add BottomBar composable
Create a `BottomBar` composable for app navigation. This component uses `NavigationBar` and `NavigationBarItem` to display navigation options for `Home`, `Carrito`, `ListaDeDeseos`, and `MiCuenta`.

The navigation logic ensures a single instance of each screen and saves the back stack state for a smoother user experience. The styling of the items changes to indicate the currently selected screen.