feat: Add ListByIdsUseCase for products
Create a new use case, `ListByIdsUseCase`, to fetch a list of `Producto` entities by their IDs from the repository.

This use case takes a list of integer IDs and returns a `Flow` that emits the corresponding list of products wrapped in a `Resource` state. It also updates the `ProductoUseCases` data class to include this new use case.